Add dropdown menu that contain link for authentication
