+++ app/views/home/_header.slim
@@ -1,4 +1,4 @@
-nav.sticky.top-0.left-0.w-screen.pt-4.bg-white.lg:px-36
+nav.sticky.top-0.left-0.w-screen.pt-4.bg-white.lg:px-36[data-controller="home"]
   .flex.items-center.justify-between.relative
     .text-pinkRed
       svg[width="102" height="32" style="display:block"]
@@ -25,21 +25,37 @@ nav.sticky.top-0.left-0.w-screen.pt-4.bg-white.lg:px-36
       a[href="#"]
         svg[viewbox="0 0 16 16" xmlns="http://www.w3.org/2000/svg" style="display: block; height: 16px; width: 16px; fill: currentcolor;" aria-hidden="true" role="presentation" focusable="false"]
           path[d="m8.002.25a7.77 7.77 0 0 1 7.748 7.776 7.75 7.75 0 0 1 -7.521 7.72l-.246.004a7.75 7.75 0 0 1 -7.73-7.513l-.003-.245a7.75 7.75 0 0 1 7.752-7.742zm1.949 8.5h-3.903c.155 2.897 1.176 5.343 1.886 5.493l.068.007c.68-.002 1.72-2.365 1.932-5.23zm4.255 0h-2.752c-.091 1.96-.53 3.783-1.188 5.076a6.257 6.257 0 0 0 3.905-4.829zm-9.661 0h-2.75a6.257 6.257 0 0 0 3.934 5.075c-.615-1.208-1.036-2.875-1.162-4.686l-.022-.39zm1.188-6.576-.115.046a6.257 6.257 0 0 0 -3.823 5.03h2.75c.085-1.83.471-3.54 1.059-4.81zm2.262-.424c-.702.002-1.784 2.512-1.947 5.5h3.904c-.156-2.903-1.178-5.343-1.892-5.494l-.065-.007zm2.28.432.023.05c.643 1.288 1.069 3.084 1.157 5.018h2.748a6.275 6.275 0 0 0 -3.929-5.068z"]
-      button.flex.items-center.gap-x-3.border.border-gray-300.rounded-full.px-3.py-2
-        . class="space-y-0.5"
-          . class="w-3.5 h-0.5 bg-gray-600"
-          . class="w-3.5 h-0.5 bg-gray-600"
-          . class="w-3.5 h-0.5 bg-gray-600"
-        .w-8.h-8.rounded-full
-          svg[viewbox="0 0 32 32" xmlns="http://www.w3.org/2000/svg" style="display: block; height: 100%; width: 100%; fill: currentcolor;" aria-hidden="true" role="presentation" focusable="false"]
-            path[d="m16 .7c-8.437 0-15.3 6.863-15.3 15.3s6.863 15.3 15.3 15.3 15.3-6.863 15.3-15.3-6.863-15.3-15.3-15.3zm0 28c-4.021 0-7.605-1.884-9.933-4.81a12.425 12.425 0 0 1 6.451-4.4 6.507 6.507 0 0 1 -3.018-5.49c0-3.584 2.916-6.5 6.5-6.5s6.5 2.916 6.5 6.5a6.513 6.513 0 0 1 -3.019 5.491 12.42 12.42 0 0 1 6.452 4.4c-2.328 2.925-5.912 4.809-9.933 4.809z"]
+      
+      
+      
+      
+      .relative.header__dropdown [data-dropdown]
+        button.header__dropdown-toggle.flex.items-center.gap-x-3.border.border-gray-300.rounded-full.px-3.py-2 [data-dropdown-button]
+          .pointer-events-none class="space-y-0.5"
+            . class="w-3.5 h-0.5 bg-gray-600"
+            . class="w-3.5 h-0.5 bg-gray-600"
+            . class="w-3.5 h-0.5 bg-gray-600"
+          .w-8.h-8.rounded-full.pointer-events-none
+            svg[viewbox="0 0 32 32" xmlns="http://www.w3.org/2000/svg" style="display: block; height: 100%; width: 100%; fill: currentcolor;" aria-hidden="true" role="presentation" focusable="false"]
+              path[d="m16 .7c-8.437 0-15.3 6.863-15.3 15.3s6.863 15.3 15.3 15.3 15.3-6.863 15.3-15.3-6.863-15.3-15.3-15.3zm0 28c-4.021 0-7.605-1.884-9.933-4.81a12.425 12.425 0 0 1 6.451-4.4 6.507 6.507 0 0 1 -3.018-5.49c0-3.584 2.916-6.5 6.5-6.5s6.5 2.916 6.5 6.5a6.513 6.513 0 0 1 -3.019 5.491 12.42 12.42 0 0 1 6.452 4.4c-2.328 2.925-5.912 4.809-9.933 4.809z"]
+        
+        ul.header__dropdown-menu.absolute.w-72.right-0.top-14.py-2.px-4.bg-white.rounded-lg.z-50.text-left
+            li.header__dropdown-item.w-full.flex.flex-col.items-start.pb-4.border-b.border-gray-300
+              a.font-bold [href="#"]
+                |Sign up
+              a[href="#"]
+                |Log in
+            li.header__dropdown-item.pt-4
+              = link_to(host_home_index_path) do
+                |Host your home
+  
   .h-16.mt-16.flex.justify-between.w-full
-    .container.relative.px-8[data-controller="tag"]
-      button.absolute.border.rounded-full.border-gray-500.p-1.5.hover:p-2.top-1/4.left-0 [data-action="click->tag#scrollLeft" type="button"]
+    .container.relative.px-8
+      button.absolute.border.rounded-full.border-gray-500.p-1.5.hover:p-2.top-1/4.left-0 [data-action="click->home#scrollLeft" type="button"]
         svg[viewbox="0 0 32 32" xmlns="http://www.w3.org/2000/svg" style="display: block; fill: none; height: 12px; width: 12px; stroke: currentcolor; stroke-width: 5.33333px; overflow: visible;" aria-hidden="true" role="presentation" focusable="false"]
           g[fill="none"]
             path[d="m20 28-11.29289322-11.2928932c-.39052429-.3905243-.39052429-1.0236893 0-1.4142136l11.29289322-11.2928932"]
-      .tags__bar.flex.gap-8.h-16.overflow-auto.whitespace-nowrap[data-tag-target="bar"]
+      .tags__bar.flex.gap-8.h-16.overflow-auto.whitespace-nowrap[data-home-target="bar"]
         - @tags.each do |tag|
           - if tag.id == 1
             input type="radio" id="tag-#{tag.id}" name="tag" value="#{tag.id}" checked=""
@@ -49,7 +65,7 @@ nav.sticky.top-0.left-0.w-screen.pt-4.bg-white.lg:px-36
             = image_tag asset_path("tag/#{tag.icon_path}"), class: "tag__image mx-auto"
             h1.my-2.text-center.font-semibold
               = tag.name
-      button.border.rounded-full.border-gray-500.p-1.5.hover:p-2.absolute.top-1/4.right-0 [data-action="click->tag#scrollRight" type="button"]
+      button.border.rounded-full.border-gray-500.p-1.5.hover:p-2.absolute.top-1/4.right-0 [data-action="click->home#scrollRight" type="button"]
         svg[viewbox="0 0 32 32" xmlns="http://www.w3.org/2000/svg" style="display: block; fill: none; height: 12px; width: 12px; stroke: currentcolor; stroke-width: 5.33333px; overflow: visible;" aria-hidden="true" role="presentation" focusable="false"]
           g[fill="none"]
             path[d="m12 4 11.2928932 11.2928932c.3905243.3905243.3905243 1.0236893 0 1.4142136l-11.2928932 11.2928932"]
